separate user routes and functions
