+++ routes/userRoutes.js
@@ -1,24 +1,12 @@
-
-const mongoose = require('mongoose');
-const User = mongoose.model('users');
+const { 
+  getUserById,
+  createUser
+} = require('./userRoutesHandlers');
 
 module.exports = (app) => {
   // Get  
-  app.get(`/api/users/:id`, async (req, res) => {
-    const {id} = req.params;
-    let user = await User.findById(id);
-    return res.status(200).send(user);
-  });
+  app.get(`/api/users/:id`, getUserById);
 
   // Create
-  app.post(`/api/users`, async (req, res) => {
-    let user = await User.findOne({email: req.body.email});
-    if (!user) {
-        user = await User.create(req.body);
-    }
-    return res.status(201).send({
-      error: false,
-      user
-    })
-  })
+  app.post(`/api/users`, createUser); 
 }
